+++ CHANGELOG.md
@@ -8,7 +8,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added TLS support to the input and output `socket` components. (@eadwright)
- 
+- Go API: New `public/schema` package for serde to and from a common schema format. (@Jeffail)
+- New Bloblang method `infer_schema`. (@Jeffail)
+
 ## 4.53.1 - 2025-07-08
 
 ### Fixed